Rename stuff from tiff to tif
As we decided to use 3 letters short form for file types.
